MLG clustering / filtering algorithm complete.
Still needs unit tests and R-level features, but mostly feature complete.

+++ src/mlg_clustering.c
@@ -47,18 +47,15 @@ SEXP farthest_neighbor(SEXP dist, SEXP mlg, SEXP threshold)
 {
 
   // TODO: More descriptive header
-  // TODO: Implement wrapper function in R
   // TODO: Make unit tests
-  // TODO: Clean up code. Get rid of commented out code.
   // TODO: Input validation
-  // TODO: Remove empty genotypes before returning.
-  // TODO: Order genotypes by size before returning?
 
   // Assumptions:
   //  dist is an n by n matrix containing distances between individuals
   //  mlg is a vector of length n containing mlg assignments
   //  threshold is a real
-  //  dist as attr "dim" of (n,n)
+  //  mlg is a vector of integers
+  //  dist is a matrix of reals
 
   int num_individuals; // n above; total number of individuals
   int num_clusters; // Number of clusters; updated upon merge
@@ -116,7 +113,7 @@ SEXP farthest_neighbor(SEXP dist, SEXP mlg, SEXP threshold)
   num_clusters = 0;
   for(int i = 0; i < num_individuals; i++)
   {
-    cur_mlg = REAL(mlg)[i]; // Get the initial cluster of this individual / Casts as int implicitly
+    cur_mlg = INTEGER(mlg)[i]; // Get the initial cluster of this individual / Casts as int implicitly
     // Insert base 1 index of individual in the next open cell for its cluster  
     cluster_matrix[cur_mlg-1][cluster_size[cur_mlg-1]] = i;
     cluster_size[cur_mlg-1]++;
@@ -161,11 +158,10 @@ SEXP farthest_neighbor(SEXP dist, SEXP mlg, SEXP threshold)
           for(int l_individual = 0; l_individual < num_individuals && cluster_matrix[j_cluster][l_individual] != -1; l_individual++)
           {
             int l_individual_val = cluster_matrix[j_cluster][l_individual];
-            // TODO: Clean up and explain the indexing here
+            // Index into the dist matrix using partial c-style indexing
             if(REAL(dist)[k_individual_val*num_individuals+l_individual_val] > tmp_max)
             {
               tmp_max = REAL(dist)[k_individual_val*num_individuals + l_individual_val];
-              //printf("tmp_max is now %f\n", tmp_max);
             }
           }
         }
@@ -173,20 +169,17 @@ SEXP farthest_neighbor(SEXP dist, SEXP mlg, SEXP threshold)
         // Make sure not to count empty clusters (tmp_max ~= -1)
         if(tmp_max > -0.5 && (tmp_max < min_cluster_distance || min_cluster_distance < 0))
         {
-          //printf("Closest pair is now: %d,%d at %f\n", i_cluster, j_cluster, tmp_max);
+          // i_cluster and j_cluster are, for now, marked for merging 
           min_cluster_distance = tmp_max;
           closest_pair[0] = i_cluster;
           closest_pair[1] = j_cluster;
         }
       }
     }
-    //printf("The closest pair is: %i and %i at %f\n", closest_pair[0], closest_pair[1], min_cluster_distance);
-
     // Merge the two closest clusters together into the lower numbered cluster
-    //  if they are withing the threshold distance from each other
+    //  if they are within the threshold distance from each other
     if(min_cluster_distance < thresh)
     {
-      
       for(int i = 0; i < cluster_size[closest_pair[1]]; i++)
       {
         cluster_matrix[closest_pair[0]][cluster_size[closest_pair[0]]] = cluster_matrix[closest_pair[1]][i];
@@ -198,16 +191,14 @@ SEXP farthest_neighbor(SEXP dist, SEXP mlg, SEXP threshold)
     }
   }
 
-  //printf("numClusters: %d\n", num_clusters);
   // Fill return vector
   for(int i = 0; i < num_mlgs; i++)
   {
     for(int j = 0; j < num_individuals && cluster_matrix[i][j] > -0.5; j++)
     {
-      //printf("%d ",cluster_matrix[i][j]);
+      // Each element of Rout should contain that individual's new mlg
       INTEGER(Rout)[cluster_matrix[i][j]] = i+1;
     }
-    //printf("\n");
   }
